kinda confused int na Long przy id w skrypcie zmiana na serial czyli BigDecimal z inkrementacja stworzona klasa abstrakcyjna by tworzyc dtosy post zwraca id no ale post dziala

+++ src/main/java/pl/pollub/inzynierka/domain/entities/AddressEntity.java
@@ -4,7 +4,6 @@ import lombok.Data;
 
 import javax.persistence.*;
 import java.util.Collection;
-import java.util.Objects;
 
 @Entity
 @Data
@@ -25,7 +24,7 @@ public class AddressEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "id", nullable = false)
-    private int id;
+    private Long id;
     @ManyToOne
     @JoinColumn(name = "city_id", referencedColumnName = "id", nullable = false)
     private CityEntity cityByCityId;
